requirements.txt
flask_sqlalchemy==3.1.1
Bootstrap_Flask==2.2.0
Flask==2.3.2
WTForms==3.0.1
Flask_WTF==1.2.1
Werkzeug==3.0.0
bootstrap-flask==2.0.1
python-dotenv==0.19.2
